Officially drop support for pre-go1.7
- Remove conditional tests on go 1.7.
- Stop running on travis with go 1.5.4.
- Bump go1.11 to go1.11.4.
