+++ experimental/devices/ina219/ina219_test.go
@@ -2,8 +2,6 @@
 // Use of this source code is governed under the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 
-// +build go1.7
-
 package ina219
 
 import (
