+++ host/sysfs/sysfs_test.go
@@ -6,6 +6,7 @@ package sysfs
 
 import (
 	"errors"
+	"io"
 
 	"periph.io/x/periph/host/fs"
 )
@@ -57,8 +58,7 @@ func (f *file) Read(p []byte) (int, error) {
 }
 
 func (f *file) Seek(offset int64, whence int) (int64, error) {
-	// io.SeekStart was added in go1.7.
-	if offset == 0 && whence == 0 {
+	if offset == 0 && whence == io.SeekStart {
 		return 0, nil
 	}
 	return 0, errors.New("not implemented")
